requirements.txt
flask~=3.0.0
werkzeug~=3.0.1
python-dotenv~=1.0.0
click~=8.1.7
pymysql~=1.1.0